refactor: messageLog-migrate-simulationLog
