+++ data/src/main/java/com/saegil/data/di/DataModule.kt
@@ -5,19 +5,16 @@ import com.saegil.data.local.TokenDataSource
 import com.saegil.data.local.TokenDataSourceImpl
 import com.saegil.data.remote.FeedService
 import com.saegil.data.remote.MapService
-import com.saegil.data.remote.MessageLogService
 import com.saegil.data.remote.OAuthService
 import com.saegil.data.remote.ScenarioService
 import com.saegil.data.remote.SimulationLogService
 import com.saegil.data.repository.FeedRepositoryImpl
 import com.saegil.data.repository.MapRepositoryImpl
-import com.saegil.data.repository.MessageLogRepositoryImpl
 import com.saegil.data.repository.OAuthRepositoryImpl
 import com.saegil.data.repository.SimulationLogRepositoryImpl
 import com.saegil.data.repository.ScenarioRepositoryImpl
 import com.saegil.domain.repository.FeedRepository
 import com.saegil.domain.repository.MapRepository
-import com.saegil.domain.repository.MessageLogRepository
 import com.saegil.domain.repository.OAuthRepository
 import com.saegil.domain.repository.ScenarioRepository
 import com.saegil.domain.repository.SimulationLogRepository
@@ -72,9 +69,4 @@ object DataModule {
         return SimulationLogRepositoryImpl(simulationLogService)
     }
 
-    @Provides
-    @Singleton
-    fun provideMessageLogRepository(messageLogService: MessageLogService): MessageLogRepository {
-        return MessageLogRepositoryImpl(messageLogService)
-    }
 }
