+++ data/src/main/java/com/saegil/data/remote/FeedServiceImpl.kt
@@ -1,7 +1,5 @@
 package com.saegil.data.remote
 
-
-import android.util.Log
 import com.saegil.data.model.NoticeListDto
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
