--- data/src/main/java/com/saegil/data/remote/MessageLogService.kt
@@ -1,9 +1,0 @@
-package com.saegil.data.remote
-
-import com.saegil.data.model.SimulationLogDetailDto
-
-interface MessageLogService {
-
-    suspend fun getMessages(simulationId: Long): SimulationLogDetailDto
-
-}
