+++ data/src/main/java/com/saegil/data/repository/SimulationLogRepositoryImpl.kt
@@ -1,7 +1,9 @@
 package com.saegil.data.repository
 
 import com.saegil.data.remote.SimulationLogService
+import com.saegil.domain.model.SimulationLogDetail
 import com.saegil.domain.repository.SimulationLogRepository
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
@@ -11,6 +13,10 @@ class SimulationLogRepositoryImpl @Inject constructor(
 
     override fun getSimulationLogs() = flow {
         emit(simulationLogService.getSimulationLogList().map { it.toDomain() })
+    }
+
+    override fun getMessages(id: Long): Flow<SimulationLogDetail> = flow {
+        emit(simulationLogService.getMessages(id).toDomain())
     }
 
 }
