+++ domain/src/main/java/com/saegil/domain/usecase/GetMessageLogUseCase.kt
@@ -1,10 +1,10 @@
 package com.saegil.domain.usecase
 
-import com.saegil.domain.repository.MessageLogRepository
+import com.saegil.domain.repository.SimulationLogRepository
 import javax.inject.Inject
 
 class GetMessageLogUseCase @Inject constructor(
-    private val messageLogRepository: MessageLogRepository
+    private val simulationLogRepository: SimulationLogRepository
 ){
-    operator fun invoke(id: Long) = messageLogRepository.getMessages(id)
+    operator fun invoke(id: Long) = simulationLogRepository.getMessages(id)
 }
